src/requirements.txt
uvicorn[standard]
requests
pipenv
curlify
pyJWT